+++ src/profile/games/cclcc/savemenu.cpp
@@ -42,7 +42,7 @@ void Configure() {
   for (size_t i = 0; i < sizeof(SaveMenuTypeNames) / sizeof(*SaveMenuTypeNames);
        i++) {
     UI::SaveMenuPageType menuType =
-        UI::SaveMenuPageType::_from_integral_unchecked(i);
+        UI::SaveMenuPageType::_from_integral_unchecked(static_cast<int>(i));
     std::string menuName = menuType._to_string();
 
     MenuTextSprite[menuType] =
