+++ src/renderer/vulkan/renderer.cpp
@@ -1814,7 +1814,7 @@ void Renderer::CaptureScreencap(Sprite& sprite) {
 int Renderer::GetSpriteSheetImage(SpriteSheet const& sheet,
                                   std::span<uint8_t> outBuffer) {
   const size_t bufferSize =
-      (size_t)(sheet.DesignWidth * sheet.DesignHeight * 4);
+      static_cast<size_t>(sheet.DesignWidth * sheet.DesignHeight * 4);
   assert(outBuffer.size() >= bufferSize);
 
   // Create a staging buffer to copy the image data to
@@ -1882,7 +1882,7 @@ int Renderer::GetSpriteSheetImage(SpriteSheet const& sheet,
 
   // Clean up the staging buffer
   vmaDestroyBuffer(Allocator, stagingBuffer.Buffer, stagingBuffer.Allocation);
-  return bufferSize;
+  return static_cast<int>(bufferSize);
 }
 
 void Renderer::EnableScissor() {}
